+++ src/main/java/org/spongepowered/common/data/builder/manipulator/mutable/entity/SneakingDataBuilder.java
@@ -1,0 +1,64 @@
+/*
+ * This file is part of Sponge, licensed under the MIT License (MIT).
+ *
+ * Copyright (c) SpongePowered <https://www.spongepowered.org>
+ * Copyright (c) contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.spongepowered.common.data.builder.manipulator.mutable.entity;
+
+import static org.spongepowered.common.data.util.DataUtil.checkDataExists;
+import static org.spongepowered.common.data.util.DataUtil.getData;
+import net.minecraft.entity.Entity;
+import org.spongepowered.api.data.DataHolder;
+import org.spongepowered.api.data.DataView;
+import org.spongepowered.api.data.key.Keys;
+import org.spongepowered.api.data.manipulator.DataManipulatorBuilder;
+import org.spongepowered.api.data.manipulator.immutable.entity.ImmutableSneakingData;
+import org.spongepowered.api.data.manipulator.mutable.entity.SneakingData;
+import org.spongepowered.api.service.persistence.InvalidDataException;
+import org.spongepowered.common.data.manipulator.mutable.entity.SpongeSneakingData;
+
+import com.google.common.base.Optional;
+
+public class SneakingDataBuilder implements DataManipulatorBuilder<SneakingData, ImmutableSneakingData> {
+
+	@Override
+	public Optional<SneakingData> build(DataView container) throws InvalidDataException {
+		checkDataExists(container, Keys.IS_SNEAKING.getQuery());
+        final boolean sneaking = getData(container, Keys.IS_SNEAKING, Boolean.class);
+        return Optional.<SneakingData>of(new SpongeSneakingData(sneaking));
+	}
+
+	@Override
+	public SneakingData create() {
+		return new SpongeSneakingData(false);
+	}
+
+	@Override
+	public Optional<SneakingData> createFrom(DataHolder dataHolder) {
+		if (dataHolder instanceof Entity) {
+            final boolean sneaking = ((Entity) dataHolder).isSneaking();
+            return Optional.<SneakingData>of(new SpongeSneakingData(sneaking));
+        }
+        return Optional.absent();
+	}
+
+}
