+++ src/main/java/org/spongepowered/common/data/processor/data/entity/SneakingDataProcessor.java
@@ -25,7 +25,6 @@
 package org.spongepowered.common.data.processor.data.entity;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static org.spongepowered.common.data.util.DataUtil.checkDataExists;
 import static org.spongepowered.common.data.util.DataUtil.getData;
 
 import com.google.common.base.Optional;
@@ -34,7 +33,6 @@ import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.data.DataHolder;
 import org.spongepowered.api.data.DataTransactionBuilder;
 import org.spongepowered.api.data.DataTransactionResult;
-import org.spongepowered.api.data.DataView;
 import org.spongepowered.api.data.key.Key;
 import org.spongepowered.api.data.key.Keys;
 import org.spongepowered.api.data.manipulator.immutable.entity.ImmutableSneakingData;
@@ -43,29 +41,11 @@ import org.spongepowered.api.data.merge.MergeFunction;
 import org.spongepowered.api.data.value.BaseValue;
 import org.spongepowered.api.data.value.immutable.ImmutableValue;
 import org.spongepowered.api.data.value.mutable.Value;
-import org.spongepowered.api.service.persistence.InvalidDataException;
-import org.spongepowered.common.data.DataProcessor;
 import org.spongepowered.common.data.manipulator.immutable.entity.ImmutableSpongeSneakingData;
 import org.spongepowered.common.data.manipulator.mutable.entity.SpongeSneakingData;
+import org.spongepowered.common.data.processor.common.AbstractSpongeDataProcessor;
 
-public class SneakingDataProcessor implements DataProcessor<SneakingData, ImmutableSneakingData> {
-
-    @Override
-    public Optional<SneakingData> build(DataView container) throws InvalidDataException {
-        checkDataExists(container, Keys.IS_SNEAKING.getQuery());
-        final boolean sneaking = getData(container, Keys.IS_SNEAKING, Boolean.class);
-        return Optional.<SneakingData>of(new SpongeSneakingData(sneaking));
-    }
-
-    @Override
-    public SneakingData create() {
-        return new SpongeSneakingData(false);
-    }
-
-    @Override
-    public ImmutableSneakingData createImmutable() {
-        return new ImmutableSpongeSneakingData(false);
-    }
+public class SneakingDataProcessor extends AbstractSpongeDataProcessor<SneakingData, ImmutableSneakingData> {
 
     @Override
     public Optional<SneakingData> createFrom(DataHolder dataHolder) {
@@ -91,17 +71,6 @@ public class SneakingDataProcessor implements DataProcessor<SneakingData, Immuta
     }
 
     @Override
-    public Optional<SneakingData> fill(DataHolder dataHolder, SneakingData manipulator) {
-        if (dataHolder instanceof Entity) {
-            final Value<Boolean> sneakingValue = manipulator.sneaking();
-            final boolean sneaking = ((Entity) dataHolder).isSneaking();
-            sneakingValue.set(sneaking);
-            return Optional.of(manipulator.set(sneakingValue));
-        }
-        return Optional.absent();
-    }
-
-    @Override
     public Optional<SneakingData> fill(DataHolder dataHolder, SneakingData manipulator, MergeFunction overlap) {
         if (dataHolder instanceof Entity) {
             final Optional<SneakingData> oldData = from(dataHolder);
@@ -116,23 +85,6 @@ public class SneakingDataProcessor implements DataProcessor<SneakingData, Immuta
     public Optional<SneakingData> fill(DataContainer container, SneakingData sneakingData) {
         sneakingData.set(Keys.IS_SNEAKING, getData(container, Keys.IS_SNEAKING));
         return Optional.of(sneakingData);
-    }
-
-    @Override
-    public DataTransactionResult set(DataHolder dataHolder, SneakingData manipulator) {
-        if (dataHolder instanceof Entity) {
-            final ImmutableValue<Boolean> newValue = manipulator.sneaking().asImmutable();
-            final SneakingData old = from(dataHolder).get();
-            final ImmutableValue<Boolean> oldValue = old.asImmutable().sneaking();
-            final boolean sneaking = manipulator.sneaking().get();
-            try {
-                ((Entity) dataHolder).setSneaking(sneaking);
-                return DataTransactionBuilder.successReplaceResult(newValue, oldValue);
-            } catch (Exception e) {
-                DataTransactionBuilder.errorResult(newValue);
-            }
-        }
-        return DataTransactionBuilder.failResult(manipulator.getValues());
     }
 
     @Override
